feat: SSR для профиля
