+++ src/app/(auth)/(withMenuNavigation)/profile/settings/page.tsx
@@ -1,16 +1,14 @@
 import type { Metadata } from 'next';
 
-import { userApi } from '@/api/api';
-import type { PersonDetailDTO } from '@/api/axios-client';
-
 import Settings from './Settings';
 
 export const metadata: Metadata = {
     title: 'Настройки профиля',
-    description: '',
+    description: 'Настройки вашего аккаунта и приложения',
 };
 
+export const dynamic = 'force-static';
+
 export default async function Page() {
-    const user: PersonDetailDTO = (await userApi.userGetPersonalDetails()).data;
-    return <Settings initUser={user} />;
+    return <Settings />;
 }
